Track Transaction Fee By Account
