+++ main.go
@@ -3,13 +3,11 @@ package main
 import (
 	"context"
 	"flag"
-	"fmt"
 	"math/big"
 	"strings"
 	"time"
 
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/ethereum/go-ethereum/consensus/misc/eip4844"
 	gethtypes "github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/ethclient"
@@ -119,7 +117,7 @@ func main() {
 				r, err := ec.TransactionReceipt(context.Background(), hash)
 				if err == nil && r.BlockHash == h.Hash() {
 					log.WithFields(txData(tx, chainID)).Infof("Transaction was included in block %d in %s", r.BlockNumber.Uint64(), time.Since(txTime[hash]))
-					recordTxInclusion(tx, chainID, time.Since(txTime[hash]))
+					recordTxInclusion(r, tx, chainID, time.Since(txTime[hash]))
 					blobsIncluded += len(tx.BlobHashes())
 					delete(pendingTxs, hash)
 					delete(txTime, hash)
@@ -168,73 +166,4 @@ func main() {
 			log.Infof("*/-------------------------------------------------------------------------------------------------------------------------------------------------------------------*/")
 		}
 	}
-}
-
-func txData(tx *gethtypes.Transaction, chainID *big.Int) log.Fields {
-	acc, err := gethtypes.Sender(gethtypes.NewCancunSigner(chainID), tx)
-	if err != nil {
-		log.WithError(err).Error("Could not get sender's account address")
-		return nil
-	}
-	accName := acc.String()
-	if name, ok := accountLabels[[20]byte(acc.Bytes())]; ok {
-		accName = name
-	}
-
-	return log.Fields{
-		"TxHash":              tx.Hash(),
-		"BlobGasFeeCap(Gwei)": float64(tx.BlobGasFeeCap().Uint64()) / params.GWei,
-		"BlobGas":             tx.BlobGas(),
-		"BlobCount":           len(tx.BlobHashes()),
-		"GasFeeCap(Gwei)":     float64(tx.GasFeeCap().Uint64()) / params.GWei,
-		"GasTipCap(Gwei)":     float64(tx.GasTipCap().Uint64()) / params.GWei,
-		"Gas":                 tx.Gas(),
-		"Account":             accName,
-	}
-}
-
-var accountLabels = map[[20]byte]string{
-	mustDecode("0xc1b634853cb333d3ad8663715b08f41a3aec47cc"): "Arbitrum",
-	mustDecode("0x6887246668a3b87f54deb3b94ba47a6f63f32985"): "Optimism",
-	mustDecode("0x5050f69a9786f081509234f1a7f4684b5e5b76c9"): "Base",
-	mustDecode("0x000000633b68f5d8d3a86593ebb815b4663bcbe0"): "Taiko",
-	mustDecode("0x2c169dfe5fbba12957bdd0ba47d9cedbfe260ca7"): "Starknet",
-	mustDecode("0x0D3250c3D5FAcb74Ac15834096397a3Ef790ec99"): "ZkSync",
-	mustDecode("0xcf2898225ed05be911d3709d9417e86e0b4cfc8f"): "Scroll",
-	mustDecode("0x415c8893d514f9bc5211d36eeda4183226b84aa7"): "Blast",
-	mustDecode("0xa9268341831efa4937537bc3e9eb36dbece83c7e"): "Linea",
-}
-
-func mustDecode(address string) [20]byte {
-	byteAddr := hexutil.MustDecode(address)
-	return [20]byte(byteAddr)
-}
-
-func recordTxMetrics(tx *gethtypes.Transaction, chainID *big.Int) {
-	acc, err := gethtypes.Sender(gethtypes.NewCancunSigner(chainID), tx)
-	if err != nil {
-		log.WithError(err).Error("Could not get sender's account address")
-		return
-	}
-	accName := acc.String()
-	if name, ok := accountLabels[[20]byte(acc.Bytes())]; ok {
-		accName = name
-	}
-	transactionsObservedGauge.WithLabelValues(accName, fmt.Sprintf("%d", len(tx.BlobHashes())), fmt.Sprintf("%d", tx.BlobGasFeeCap().Uint64())).Inc()
-}
-
-func recordTxInclusion(tx *gethtypes.Transaction, chainID *big.Int, inclusionDelay time.Duration) {
-	acc, err := gethtypes.Sender(gethtypes.NewCancunSigner(chainID), tx)
-	if err != nil {
-		log.WithError(err).Error("Could not get sender's account address")
-		return
-	}
-	accName := acc.String()
-	if name, ok := accountLabels[[20]byte(acc.Bytes())]; ok {
-		accName = name
-	}
-	gasTip, _ := tx.GasTipCap().Float64()
-	gasTipGwei := gasTip / params.GWei
-	transactionInclusionDelay.WithLabelValues(accName, fmt.Sprintf("%d", len(tx.BlobHashes()))).Observe(inclusionDelay.Seconds())
-	transactionTip.WithLabelValues(accName, fmt.Sprintf("%d", len(tx.BlobHashes()))).Observe(gasTipGwei)
 }
